examples/web_agent/pyproject.toml
[tool.poetry]
name = "airtop-webagent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.26.1"
requests = "^2.32.3"
airtop = "^0.0.24"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
